--- package.json
@@ -1,18 +1,0 @@
-{
-  "name": "fullstackfour",
-  "version": "1.0.0",
-  "description": "(rough notes)",
-  "main": "app.js",
-  "scripts": {
-    "test": "echo \"Error: no test specified\" && exit 1"
-  },
-  "keywords": [],
-  "author": "",
-  "license": "ISC",
-  "dependencies": {
-    "body-parser": "^1.20.2",
-    "cors": "^2.8.5",
-    "dotenv": "^16.4.5",
-    "mysql2": "^3.11.0"
-  }
-}
